requirements.txt
attrs==22.1.0
greenlet==1.1.2
iniconfig==1.1.1
msgpack==1.0.3
mypy==0.971
mypy-extensions==0.4.3
neovim==0.3.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pynvim==0.4.3
pyparsing==3.0.9
pytest==7.1.2
PyYAML==6.0
tomli==2.0.1
types-PyYAML==6.0.11
typing_extensions==4.3.0
